test: `Cell.floor`

+++ src/safeds/data/tabular/containers/_lazy_cell.py
@@ -88,7 +88,10 @@ class _LazyCell(Cell[T]):
         return _wrap(self._expression.ceil())
 
     def __floor__(self) -> Cell[R]:
-        return _wrap(self._expression.floor())
+        import polars as pl
+
+        # polars does not yet implement floor for integers
+        return _wrap(self._expression.cast(pl.Float64).floor())
 
     def __neg__(self) -> Cell[R]:
         return _wrap(self._expression.__neg__())
